[Refactor] Review: 패키지 구조 수정 및 docs 생성 #205
