+++ review/src/main/java/com/haot/review/application/service/ReviewServiceImpl.java
@@ -9,8 +9,8 @@ import com.haot.review.common.response.enums.ErrorCode;
 import com.haot.review.domain.model.Review;
 import com.haot.review.domain.model.ReviewImage;
 import com.haot.review.domain.repository.ReviewRepository;
-import com.haot.review.presentation.client.LodgeClient;
-import com.haot.review.presentation.dto.LodgeReadOneResponse;
+import com.haot.review.infrastructure.client.LodgeClient;
+import com.haot.review.infrastructure.dto.LodgeReadOneResponse;
 import com.haot.submodule.role.Role;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
