+++ review/src/main/java/com/haot/review/infrastructure/client/LodgeClient.java
@@ -1,7 +1,7 @@
-package com.haot.review.presentation.client;
+package com.haot.review.infrastructure.client;
 
 import com.haot.review.common.response.ApiResponse;
-import com.haot.review.presentation.dto.LodgeReadOneResponse;
+import com.haot.review.infrastructure.dto.LodgeReadOneResponse;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
