+++ review/src/main/java/com/haot/review/infrastructure/dto/LodgeReadOneResponse.java
@@ -1,4 +1,4 @@
-package com.haot.review.presentation.dto;
+package com.haot.review.infrastructure.dto;
 
 import java.util.List;
