+++ review/src/main/java/com/haot/review/presentation/controller/ReviewController.java
@@ -1,6 +1,5 @@
 package com.haot.review.presentation.controller;
 
-
 import com.haot.review.application.dtos.req.ReviewCreateRequest;
 import com.haot.review.application.dtos.req.ReviewSearchRequest;
 import com.haot.review.application.dtos.req.ReviewUpdateRequest;
@@ -28,15 +27,13 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 
-@Tag(name = "Review Management", description = "리뷰 API")
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("/api/v1/reviews")
-public class ReviewController {
+public class ReviewController implements ReviewControllerDocs {
 
   private final ReviewService reviewService;
 
-  @Operation(summary = "리뷰 생성", description = "사용자가 리뷰를 작성할 수 있습니다.")
   @RoleCheck(Role.USER)
   @ResponseStatus(HttpStatus.CREATED)
   @PostMapping
@@ -48,7 +45,6 @@ public class ReviewController {
     return ApiResponse.success(reviewService.createReview(userId, request));
   }
 
-  @Operation(summary = "리뷰 단건 조회", description = "사용자가 리뷰를 조회할 수 있습니다.")
   @ResponseStatus(HttpStatus.OK)
   @GetMapping("/{reviewId}")
   public ApiResponse<ReviewGetResponse> readOne(
@@ -57,7 +53,6 @@ public class ReviewController {
     return ApiResponse.success(reviewService.readOne(reviewId));
   }
 
-  @Operation(summary = "리뷰 검색 조회", description = "사용자가 검색 조건에 따라 리뷰를 조회할 수 있습니다.")
   @RoleCheck({Role.ADMIN, Role.HOST, Role.USER})
   @ResponseStatus(HttpStatus.OK)
   @GetMapping
@@ -69,7 +64,6 @@ public class ReviewController {
     return ApiResponse.success(reviewService.searchReview(role, request, pageable));
   }
 
-  @Operation(summary = "리뷰 수정", description = "사용자가 리뷰를 수정할 수 있습니다.")
   @RoleCheck(Role.USER)
   @ResponseStatus(HttpStatus.OK)
   @PatchMapping("/{reviewId}")
@@ -83,8 +77,6 @@ public class ReviewController {
     return ApiResponse.success(SuccessCode.UPDATE_REVIEW_SUCCESS);
   }
 
-
-  @Operation(summary = "리뷰 삭제", description = "사용자가 리뷰를 삭제할 수 있습니다.")
   @RoleCheck({Role.ADMIN, Role.HOST, Role.USER})
   @ResponseStatus(HttpStatus.OK)
   @DeleteMapping("/{reviewId}")
